content-container: move folder
- from containers to components
