+++ webapp/src/components/Content/renderPalm.js
